+++ .env.example
@@ -6,4 +6,4 @@ DB_HOST="127.0.0.1"
 DB_PORT="3306"
 DB_NAME="laravel"
 DB_USER="root"
-DB_PASS="root"
+DB_PASS=""
