+++ app/Controllers/HomeController.php
@@ -1,0 +1,7 @@
+<?php 
+
+namespace App\Controller;
+
+class HomeController{
+    
+}
