+++ app/Core/Core.php
@@ -1,0 +1,6 @@
+<?php 
+
+namespace App\Core;
+
+require_once(__DIR__.'/app/Core/Include/SessionInit');
+require_once(__DIR__.'/app/Core/Include/ExceptionInit');
