+++ app/Core/Helper/Init.php
@@ -1,0 +1,7 @@
+<?php 
+
+namespace App\Core\Helper\Init;
+
+class Init{
+    
+}
