+++ app/Core/Include/ExceptionInit.php
@@ -1,0 +1,7 @@
+<?php 
+
+    // Safe Error Exceptions
+    function exception_error_handler($errno, $errstr, $errfile, $errline ) {
+        throw new ErrorException($errstr, $errno, 0, $errfile, $errline);
+    }
+    set_error_handler("exception_error_handler");
