+++ app/Core/Include/SessionInit.php
@@ -1,0 +1,4 @@
+<?php 
+
+    // START SESSIONS
+    session_start();
