+++ app/Kernel.php
@@ -1,6 +1,20 @@
 <?php
 
-namespace Framework;
+namespace App;
+
+// Core Enginee
+require_once(__DIR__.'/app/Core/Core.php');
+
+// Services
+
+        // Connect
+        require_once __DIR__ . '/vendor/autoload.php';  // Composer Packages
+        require_once __DIR__ . '/app/Kernel.php';       // Core connector
+        require_once __DIR__ . '/app/Helpers/App.php';  // Helpers
+        require_once __DIR__ . '/core.php';             // Core Loader
+
+// Helpers
+
 
 /*require_once __DIR__ . '/Core/Route.php';
 require_once __DIR__ . '/Core/Database.php';
