--- core.php
@@ -1,4 +1,0 @@
-<?php
-
-
-
